pyproject.toml
[project]
name = "ai-team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.9.3",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "mem0ai>=0.1.99",
    "qdrant-client>=1.9.0",
]
